+++ BFB/BFB.Engine/Inventory/IInventoryManager.cs
@@ -9,7 +9,7 @@ namespace BFB.Engine.Inventory
         /// Gets the active slot id
         /// </summary>
         /// <returns></returns>
-        int GetActiveSlotId();
+        byte GetActiveSlotId();
         
         /// <summary>
         /// Increments the Active slot by one
@@ -25,7 +25,7 @@ namespace BFB.Engine.Inventory
         /// Moves the active slot to the given index
         /// </summary>
         /// <param name="slotId">The slot index</param>
-        void MoveActiveSlot(int slotId);
+        void MoveActiveSlot(byte slotId);
 
         /// <summary>
         /// Gets the selected item
@@ -37,7 +37,7 @@ namespace BFB.Engine.Inventory
         /// Gets the maximum size of the inventory
         /// </summary>
         /// <returns></returns>
-        int MaxInventorySize();
+        byte MaxInventorySize();
         
         /// <summary>
         /// Indicates if the inventory is full or not
@@ -50,7 +50,7 @@ namespace BFB.Engine.Inventory
         /// </summary>
         /// <param name="slotId">The slot id to check</param>
         /// <returns>A boolean indicating if the stack is full</returns>
-        bool IsSlotFull(int slotId);
+        bool IsSlotFull(byte slotId);
         
         /// <summary>
         /// Inserts a new item into the first available spot in the inventory
@@ -65,14 +65,14 @@ namespace BFB.Engine.Inventory
         /// <param name="slotId"></param>
         /// <param name="item"></param>
         /// <returns></returns>
-        bool InsertAtSlot(int slotId, IItem item);
+        bool InsertAtSlot(byte slotId, IItem item);
         
         /// <summary>
         /// Removes the items at a given slot Id
         /// </summary>
         /// <param name="slotId">The slot to remove from</param>
         /// <returns>The items that were at that slot or null if it was empty</returns>
-        IItem Remove(int slotId);
+        IItem Remove(byte slotId);
             
         /// <summary>
         /// Inserts an item at a given location and returns the previous item at the position
@@ -80,14 +80,14 @@ namespace BFB.Engine.Inventory
         /// <param name="slotId">The slot to swap on</param>
         /// <param name="item">The item to insert</param>
         /// <returns>The item that was in the slot previously</returns>
-        IItem Swap(int slotId, IItem item);
+        IItem Swap(byte slotId, IItem item);
             
         /// <summary>
         /// Splits the stack of items at a slot in two. If there is only one item it is removed and returned
         /// </summary>
         /// <param name="slotId"></param>
         /// <returns></returns>
-        IItem Split(int slotId);
+        IItem Split(byte slotId);
 
         /// <summary>
         /// Attempts to merge two items together if the types match and gives back overflow
@@ -95,35 +95,35 @@ namespace BFB.Engine.Inventory
         /// <param name="slotId">The slot to merge</param>
         /// <param name="item">The item to merge</param>
         /// <returns>Any items that did not fit in the stack</returns>
-        IItem Merge(int slotId, IItem item);
+        IItem Merge(byte slotId, IItem item);
             
         /// <summary>
         /// Gets the item at a given slot
         /// </summary>
         /// <param name="slotId">The slot to get</param>
         /// <returns>The items at the spot</returns>
-        IItem GetSlot(int slotId);
+        IItem GetSlot(byte slotId);
         
         /// <summary>
         /// Indicates if the slot is occupied or not. Does not indicate of the slot is full.
         /// </summary>
         /// <param name="slotId">The slot to check</param>
         /// <returns>A boolean indicating if the slot is occupied</returns>
-        bool SlotOccupied(int slotId);
+        bool SlotOccupied(byte slotId);
 
         /// <summary>
         /// Indicates if the slotId Is valid
         /// </summary>
         /// <param name="slotId">The slotId to check</param>
         /// <returns>A boolean indicating if the slot is a valid range</returns>
-        bool SlotInRange(int slotId);
+        bool SlotInRange(byte slotId);
 
         /// <summary>
         /// Finds a slot with the type of item
         /// </summary>
         /// <param name="itemType">The item type</param>
         /// <returns>Null if not slots found or the slot number to match</returns>
-        int? SlotWithItemTypeAvailable(string itemType);
+        byte? SlotWithItemTypeAvailable(string itemType);
 
         /// <summary>
         /// Clears the inventory
@@ -134,7 +134,7 @@ namespace BFB.Engine.Inventory
         /// Gets all items as a list from the inventory
         /// </summary>
         /// <returns>A list of all items in the inventory</returns>
-        List<KeyValuePair<int, IItem>> GetAllItems();
+        List<KeyValuePair<byte, IItem>> GetAllItems();
 
     }
